app/requirements.txt
py-color-log>=1.0.2
flask>=2.0.3
pyyaml-erb>=1.0.2
uwsgi